chore: organize components into folders to clean up src/components/svelte
