+++ src/components/svelte/modal/ScheduleModal.svelte
@@ -1,7 +1,7 @@
 <script>
-  import { getAppData } from "../../lib/context";
-  import { queryStore } from "../../lib/store.svelte";
-  import ScheduleRender from "./ScheduleRender.svelte";
+  import { getAppData } from "../../../lib/context";
+  import { queryStore } from "../../../lib/store.svelte";
+  import ScheduleRender from "../room/ScheduleRender.svelte";
   const { classesMap } = getAppData();
   const classes = $derived(classesMap.get(queryStore.value) ?? []);
 </script>
